pyproject.toml
[project]
name = "dUltra"
version = "0.1.0"
description = "dUltra: Ultra-Fast Diffusion Large Language Models via Group Relative Policy Optimization (GRPO)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.0",
    "bitsandbytes>=0.45.3",
    "datasets==4.0.0", # In order to use APPS dataset on huggingface, downgrade to 3.6.0
    "deepspeed>=0.16.4",
    "dill",
    "evaluate>=0.4.5",
    "flash-attn==2.8.3",
    "hydra-core>=1.3.2",
    "jsonlines",
    "liger-kernel>=0.6.2",
    "math-verify>=0.8.0",
    "more_itertools",
    "numpy>=1.24",
    "openai>=2.7.1",
    "peft>=0.17.0",
    "pre-commit>=4.3.0",
    "pybind11>=2.6.2",
    "pytablewriter",
    "regex>=2025.9.1",
    "rouge-score>=0.0.4",
    "sacrebleu>=1.5.0",
    "scikit-learn>=0.24.1",
    "sentencepiece>=0.2.1",
    "setuptools>=80.9.0",
    "sqlitedict",
    "tiktoken>=0.9.0",
    "torch>=2.8.0",
    "torchaudio>=2.6.0",
    "tqdm>=4.67.1",
    "tqdm-multiprocess",
    "transformers>=4.53.2",
    "trl==0.23.0",
    "wandb>=0.21.3",
    "word2number",
    "zstandard",
    "lm-eval>=0.4.9",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]

[project.scripts]
lm-eval = "lm_eval.__main__:cli_evaluate"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["model", "SFT", "diffu_grpo", "lm_eval*"]

[tool.setuptools.package-data]
lm_eval = ["**/*.yaml", "tasks/**/*"]

[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]

[tool.pymarkdown]
plugins.md013.enabled = false
plugins.md024.allow_different_nesting = true
plugins.md025.enabled = false
plugins.md028.enabled = false
plugins.md029.allow_extended_start_values = true
plugins.md034.enabled = false

[tool.ruff.lint]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["F401", "F402", "F403"]
"utils.py" = ["F401"]
